Improve code Change license

+++ src/Corsinvest.ProxmoxVE.AutoSnap.Api/ResultSnapVm.cs
@@ -1,13 +1,6 @@
 ﻿/*
- * This file is part of the cv4pve-autosnap https://github.com/Corsinvest/cv4pve-autosnap,
- *
- * This source file is available under two different licenses:
- * - GNU General Public License version 3 (GPLv3)
- * - Corsinvest Enterprise License (CEL)
- * Full copyright and license information is available in
- * LICENSE.md which is distributed with this source code.
- *
- * Copyright (C) 2016 Corsinvest Srl	GPLv3 and CEL
+ * SPDX-License-Identifier: GPL-3.0-only
+ * SPDX-FileCopyrightText: 2020 Copyright Corsinvest Srl
  */
 
 namespace Corsinvest.ProxmoxVE.AutoSnap.Api
@@ -21,6 +14,6 @@ namespace Corsinvest.ProxmoxVE.AutoSnap.Api
         /// Vm id
         /// </summary>
         /// <value></value>
-        public int VmId { get; internal set; }
+        public long VmId { get; internal set; }
     }
 }
